requirements.txt
# garbage265 垃圾分类 — 项目依赖
# 安装: pip install -r requirements.txt

# 核心框架
torch>=1.10.0
torchvision>=0.11.0

# 图像处理
Pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0

# 数据处理 & 可视化
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0

# 数据集下载 (ModelScope)
modelscope[dataset]>=1.0.0
datasets>=3.0.0

# YOLOv5 额外依赖
PyYAML>=5.4
scipy>=1.7.0
